requirements.txt
# --- Core web server (FastAPI + SSE shim) ---
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0

# --- Config & HTTP clients ---
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
httpx>=0.27,<1.0

# --- Telegram bot ---
python-telegram-bot==21.*

# --- Database (PostgreSQL) ---
psycopg2-binary>=2.9,<3.0

# --- Models / IDs ---
pydantic>=2.10,<3.0
python-ulid>=3.0,<4.0

# --- Sentient compatibility (keep even though we use the shim) ---
sentient-agent-framework==0.3.0
